Added BMP images for water level meter and cleaned up display
There are now 5 BMP images for the water level. One is the background vase, the others are the 4 water levels that can be colored in blue (Wet, Damp, Moist, Dry). Background is now white, and the moisture level text, humidity, and temperature text have been aranged to clean up the design for readability and general appeal.