Added buttons for moisture categories and changed moisture functionality
Very Low, Low, Moderate, and High moisture category buttons have been added to the screen and remember what was pressed for each plant. If selected, it has a green background. Functionality for determining if a plant is Wet, Damp, Moist, or Dry is almost fully updated. It now is mapped to a percentage from 0-100. The percentages are based on Acurite's suggested plant moisture levels for 4 different plant categories.